+++ modules/global/src/com/egorl/car_service/service/ManagerService.java
@@ -5,5 +5,5 @@ import com.haulmont.cuba.security.entity.User;
 public interface ManagerService {
     String NAME = "carservice_ManagerService";
 
-    Integer getPurchaseRequestsCount(User user);
+    Long getPurchaseRequestsCount(User user);
 }
